refactor(trusty-search): split reindex_engine into focused submodules (closes #571)
Split the oversized `commands/reindex_ui.rs` (820 SLOC, over 500 cap) into
a focused `reindex_ui/` module directory:

- `phase.rs`   — `ReindexPhase` enum + `phase_to_bar_slot` (110 raw lines)
- `bars.rs`    — `BarState`, `bar_style`, `STAGE_LABELS`, `EMBED_STAR_NOTE`,
                 `ReindexUi` struct + impl (468 raw lines)
- `timings.rs` — `ReindexTimings`, `format_timing_breakdown`,
                 `print_timing_breakdown` (175 raw lines)
- `tests.rs`   — all unit tests extracted from the inline `mod tests` block
                 (762 raw lines; test cap = 1500 SLOC)
- `mod.rs`     — thin re-export facade; all external `use` paths unchanged

No runtime behaviour, public APIs, CLI flags, or output formatting changed.
The allowlist entry for `reindex_ui.rs` is removed; check_line_cap exits 0.
