+++ .line-cap-allowlist.tsv
@@ -73,8 +73,7 @@ crates/trusty-mpm/src/tui/dashboard.rs	696
 crates/trusty-mpm/src/tui/health.rs	2200
 crates/trusty-mpm/src/tui/mod.rs	520
 crates/trusty-search/src/commands/doctor_checks.rs	503
-crates/trusty-search/src/commands/reindex_ui.rs	820
-crates/trusty-search/src/commands/start.rs	1085
+crates/trusty-search/src/commands/start.rs	1083
 crates/trusty-search/src/core/chunker/mod.rs	1534
 crates/trusty-search/src/core/indexer/tests.rs	2472
 crates/trusty-search/src/service/call_chain.rs	738
